[ANDROSDK-592] Let Dagger instantiate RelationshipCollectionRepositoryImpl
