+++ core/src/main/java/org/hisp/dhis/android/core/relationship/RelationshipPackageDIModule.java
@@ -37,6 +37,7 @@ import retrofit2.Retrofit;
 
 @Module(includes = {
         RelationshipEntityDIModule.class,
+        RelationshipItemEntityDIModule.class,
         RelationshipTypeEntityDIModule.class
 })
 public final class RelationshipPackageDIModule {
